+++ CHANGELOG.md
@@ -6,6 +6,13 @@ dev-master
 
 ### Features
 
+- [query] Full support for manipulating multivalue properties
+
+alpha-5
+-------
+
+### Features
+
 - [shell] Added "shell:clear" command to support clearing the console output
 - [general] The shell supports being embedded as a dependency
 - [node:edit] New command `node:edit` enables editing of entire node